epay_back/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==3.2
django-admin-interface==0.26.0
django-background-tasks==1.2.5
django-colorfield==0.9.0
django-compat==1.0.15
django-cors-headers==4.1.0
idna==3.4
isodate==0.6.1
kavenegar==1.1.2
Khayyam==3.0.17
lxml==4.9.2
Pillow==9.5.0
platformdirs==3.8.0
python-slugify==8.0.1
pytz==2023.3
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
urllib3==2.0.3
zeep==4.2.1
